requirements.txt
aiofiles==23.2.1
aiogram==3.5.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
degiro-connector==3.0.18
fastapi==0.110.3
frozenlist==1.4.1
h11==0.14.0
idna==3.7
isodate==0.6.1
magic-filter==1.0.12
multidict==6.0.5
onetimepass==1.0.1
orjson==3.10.2
polars==0.20.23
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
whatsapp-python==3.0.6
wrapt==1.16.0
yarl==1.9.4
